Registration design + logic implemented

+++ app/src/main/java/com/example/bazaar_marketplace/models/Login.kt
@@ -6,7 +6,7 @@ data class LoginResponse(
     val username: String,
     val email: String,
     @SerializedName("phone_number")
-    val phoneNumber: Int,
+    val phoneNumber: Long,
     val token: String,
     @SerializedName("creation_time")
     val creationTime: Long,
